requirements.txt
moviepy==1.0.3
pandas==2.0.3
numpy==1.24.3
tqdm==4.66.1
Pillow==9.5.0
fastapi>=0.109.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
google-auth>=2.27.0
uvicorn>=0.27.0
python-multipart>=0.0.9
anthropic>=0.8.1